refactor: flatten folder structure - remove layered architecture
## Changes

### Folder Restructuring
**Removed layered architecture:**
- Deleted: `api/`, `domain/`, `infrastructure/` folders
- No more nested folder hierarchies

**Implemented flat structure:**
- `controllers/` - All API controllers (from api/controllers/)
- `middleware/` - All middleware (from api/middleware/)
- `decorators/` - All decorators (from api/decorators/)
- `services/` - All business logic (merged from domain/services/ + services/)
- `config/` - Configuration files (all auth, db, logging, email, redis, r2, queues)

### Config Files Merged
**Created `config/` with:**
- `auth.ts` - Merged better-auth.ts + jwt.ts + password.ts
- `db.ts` - Merged client.ts + schema.ts
- `logging.ts` - Moved logger.ts
- `email.ts` - Moved email.service.ts
- `redis.ts` - Moved redis.ts
- `r2.ts` - Moved r2.ts
- `queues.ts` - Moved messageRetryQueue.ts

### File Moves
| From | To |
|------|-----|
| `api/controllers/*` | `controllers/*` |
| `api/middleware/*` | `middleware/*` |
| `api/decorators/*` | `decorators/*` |
| `domain/services/*` + `services/*` | `services/*` (merged) |
| `infrastructure/auth/*` | `config/auth.ts` |
| `infrastructure/db/*` | `config/db.ts` |
| `infrastructure/logging/*` | `config/logging.ts` |
| `infrastructure/email/*` | `config/email.ts` |
| `infrastructure/redis.ts` | `config/redis.ts` |
| `infrastructure/r2.ts` | `config/r2.ts` |
| `infrastructure/queues/*` | `config/queues.ts` |

### Import Updates
- Updated all imports to use flat structure paths
- Changed from `@yacc/backend/infrastructure/*` to `../config/*`
- Changed from `@yacc/backend/domain/*` to `../services/*`

### Cleanup
- Simplified index.ts (removed unused apiRouter, Router import, runMigrations)
- Removed old layered architecture folders
- Kept: `connectors/`, `websockets/`, `workers/`, `types/`, `utils/`, `config/`

### Benefits
- Simpler file organization
- No complex folder navigation
- Easier to find files
- Follows project's flat structure preference
- All config files grouped in `config/`

## New Structure
```
packages/backend/src/
├── controllers/       ← API endpoints
├── middleware/        ← Request/response processing
├── decorators/        ← Custom decorators
├── services/          ← Business logic
├── config/            ← Configuration files
├── connectors/        ← Platform connectors
├── websockets/         ← WebSocket
├── workers/           ← Background workers
├── types/             ← Type definitions
├── utils/             ← Utility functions
└── index.ts          ← Entry point
```
