+++ packages/backend/src/config/db.ts
@@ -1,3 +1,29 @@
+import { drizzle } from 'drizzle-orm/node-postgres';
+import pkg from 'pg';
+import * as schema from './schema.js';
+import { config } from '../../config/config';
+
+const { Pool } = pkg;
+
+const pool = new Pool({
+  connectionString: config.database.url,
+});
+
+export const db = drizzle(pool, { schema });
+
+// Health check
+export async function checkDatabaseConnection(): Promise<boolean> {
+  try {
+    const result = await pool.query('SELECT NOW()');
+    console.log('✅ Database connected:', result.rows[0]);
+    return true;
+  } catch (error) {
+    console.error('❌ Database connection failed:', error);
+    return false;
+  }
+}
+
+export default db;
 import {
   pgTable,
   serial,
