--- packages/backend/src/infrastructure/auth/jwt.ts
@@ -1,75 +1,0 @@
-import jwt from 'jsonwebtoken';
-
-// CRITICAL: Production validation for JWT secret (TD-002 from GOV-008)
-const jwtSecret = process.env.JWT_SECRET;
-
-if (!jwtSecret) {
-  if (process.env.NODE_ENV === 'production') {
-    throw new Error(
-      'FATAL: JWT_SECRET is required in production. ' +
-      'Set JWT_SECRET environment variable before starting the server.'
-    );
-  }
-  console.warn(
-    '⚠️  WARNING: Using default JWT secret for development. ' +
-    'Set JWT_SECRET environment variable in production!'
-  );
-}
-
-const JWT_SECRET = jwtSecret || 'your-secret-key-change-in-production';
-const JWT_EXPIRES_IN = process.env.JWT_EXPIRES_IN || '7d';
-
-export interface JWTPayload {
-  userId: string;
-  email: string;
-  role: string;
-  iat?: number;
-  exp?: number;
-}
-
-/**
- * Sign a JWT token
- */
-export function signToken(payload: Omit<JWTPayload, 'iat' | 'exp'>): string {
-  return jwt.sign(payload, JWT_SECRET, {
-    expiresIn: JWT_EXPIRES_IN,
-  });
-}
-
-/**
- * Verify a JWT token
- */
-export function verifyToken(token: string): JWTPayload {
-  try {
-    return jwt.verify(token, JWT_SECRET) as JWTPayload;
-  } catch (error) {
-    throw new Error('Invalid or expired token');
-  }
-}
-
-/**
- * Extract token from Authorization header
- */
-export function extractToken(authHeader?: string): string | null {
-  if (!authHeader) {
-    return null;
-  }
-
-  const parts = authHeader.split(' ');
-  if (parts.length !== 2 || parts[0] !== 'Bearer') {
-    return null;
-  }
-
-  return parts[1];
-}
-
-/**
- * Decode token without verification (for debugging)
- */
-export function decodeToken(token: string): JWTPayload | null {
-  try {
-    return jwt.decode(token) as JWTPayload;
-  } catch {
-    return null;
-  }
-}
