+++ packages/backend/src/middleware/auth-betterauth.middleware.ts
